+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_params.py
@@ -16,6 +16,8 @@ from azure.mgmt.sql.models import (
     ElasticPoolPerDatabaseSettings,
     ImportExtensionRequest,
     ExportRequest,
+    ManagedDatabase,
+    ManagedInstance,
     Server,
     ServerAzureADAdministrator,
     Sku
@@ -54,7 +56,9 @@ from .custom import (
 
 from ._validators import (
     create_args_for_complex_type,
-    validate_elastic_pool_id
+    validate_elastic_pool_id,
+    validate_managed_instance_storage_size,
+    validate_subnet
 )
 
 
@@ -82,7 +86,7 @@ class SizeWithUnitConverter(object):  # pylint: disable=too-few-public-methods
         try:
             uvals = (self.unit_map[unit_part] if unit_part else 1) / \
                 (self.unit_map[self.unit] if self.unit else 1)
-            return self.result_type(uvals) * self.result_type(numeric_part)
+            return self.result_type(uvals * self.result_type(numeric_part))
         except KeyError:
             raise ValueError()
 
@@ -140,6 +144,19 @@ zone_redundant_param_type = CLIArgumentType(
     help='Specifies whether to enable zone redundancy',
     arg_type=get_three_state_flag())
 
+managed_instance_param_type = CLIArgumentType(
+    options_list=['--managed-instance', '--mi'],
+    help='Name of the Azure SQL managed instance.')
+
+storage_param_type = CLIArgumentType(
+    options_list=['--storage'],
+    type=SizeWithUnitConverter('GB', result_type=int, unit_map=dict(B=1.0 / (1024 * 1024 * 1024),
+                                                                    kB=1.0 / (1024 * 1024),
+                                                                    MB=1.0 / 1024,
+                                                                    GB=1,
+                                                                    TB=1024)),
+    help='The storage size. If no unit is specified, defaults to gigabytes (GB).',
+    validator=validate_managed_instance_storage_size)
 
 db_service_objective_examples = 'Basic, S0, P1, GP_Gen4_1, BC_Gen5_2.'
 dw_service_objective_examples = 'DW100, DW1000c'
@@ -1005,3 +1022,161 @@ def load_arguments(self, _):
         c.extra('vnet_name',
                 options_list=['--vnet-name'],
                 help='The virtual network name')
+
+    ###############################################
+    #                sql managed instance         #
+    ###############################################
+    with self.argument_context('sql mi') as c:
+        c.argument('managed_instance_name',
+                   help='The managed instance name',
+                   options_list=['--name', '-n'],
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
+
+        c.argument('tier',
+                   arg_type=tier_param_type,
+                   help='The edition component of the sku. Allowed value is GeneralPurpose.')
+
+        c.argument('family',
+                   arg_type=family_param_type,
+                   help='The compute generation component of the sku. '
+                   'Allowed values include: Gen4, Gen5.')
+
+        c.argument('storage_size_in_gb',
+                   options_list=['--storage'],
+                   arg_type=storage_param_type,
+                   help='The storage size of the managed instance. '
+                   'Storage size must be specified in increments of 32 GB')
+
+        c.argument('license_type',
+                   arg_type=get_enum_type(DatabaseLicenseType),
+                   help='The license type to apply for this managed instance.')
+
+        c.argument('vcores',
+                   options_list=['--capacity', '-c'],
+                   help='The capacity of the managed instance in vcores.')
+
+    with self.argument_context('sql mi create') as c:
+        # Create args that will be used to build up the ManagedInstance object
+        create_args_for_complex_type(
+            c, 'parameters', ManagedInstance, [
+                'administrator_login',
+                'administrator_login_password',
+                'license_type',
+                'virtual_network_subnet_id',
+                'vcores',
+                'storage_size_in_gb'
+            ])
+
+        # Create args that will be used to build up the Managed Instance's Sku object
+        create_args_for_complex_type(
+            c, 'sku', Sku, [
+                'family',
+                'name',
+                'tier',
+            ])
+
+        c.ignore('name')  # Hide sku name
+
+        c.argument('administrator_login',
+                   options_list=['--admin-user', '-u'],
+                   required=True)
+
+        c.argument('administrator_login_password',
+                   options_list=['--admin-password', '-p'],
+                   required=True)
+
+        c.extra('vnet_name',
+                options_list=['--vnet-name'],
+                help='The virtual network name',
+                validator=validate_subnet)
+
+        c.argument('virtual_network_subnet_id',
+                   options_list=['--subnet'],
+                   required=True,
+                   help='Name or ID of the subnet that allows access to an Azure Sql Managed Instance. '
+                   'If subnet name is provided, --vnet-name must be provided.')
+
+        c.argument('assign_identity',
+                   options_list=['--assign-identity', '-i'],
+                   help='Generate and assign an Azure Active Directory Identity for this managed instance '
+                   'for use with key management services like Azure KeyVault.')
+
+    with self.argument_context('sql mi update') as c:
+        # Create args that will be used to build up the ManagedInstance object
+        create_args_for_complex_type(
+            c, 'parameters', ManagedInstance, [
+                'administrator_login_password',
+            ])
+
+        c.argument('administrator_login_password',
+                   options_list=['--admin-password', '-p'])
+
+        c.argument('assign_identity',
+                   options_list=['--assign-identity', '-i'],
+                   help='Generate and assign an Azure Active Directory Identity for this managed instance '
+                   'for use with key management services like Azure KeyVault. '
+                   'If identity is already assigned - do nothing.')
+
+    ###############################################
+    #                sql managed db               #
+    ###############################################
+    with self.argument_context('sql midb') as c:
+        c.argument('managed_instance_name',
+                   arg_type=managed_instance_param_type,
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
+
+        c.argument('database_name',
+                   options_list=['--name', '-n'],
+                   help='The name of the Azure SQL Managed Database.',
+                   # Allow --ids command line argument. id_part=child_name_1 is 2nd name in uri
+                   id_part='child_name_1')
+
+    with self.argument_context('sql midb create') as c:
+        create_args_for_complex_type(
+            c, 'parameters', ManagedDatabase, [
+                'collation',
+            ])
+
+        c.argument('collation',
+                   required=False,
+                   help='The collation of the Azure SQL Managed Database collation to use, '
+                   'e.g.: SQL_Latin1_General_CP1_CI_AS or Latin1_General_100_CS_AS_SC')
+
+    with self.argument_context('sql midb restore') as c:
+        create_args_for_complex_type(
+            c, 'parameters', ManagedDatabase, [
+                'target_managed_database_name',
+                'target_managed_instance_name',
+                'restore_point_in_time'
+            ])
+
+        c.argument('target_managed_database_name',
+                   options_list=['--dest-name'],
+                   required=True,
+                   help='Name of the managed database that will be created as the restore destination.')
+
+        c.argument('target_managed_instance_name',
+                   options_list=['--dest-mi'],
+                   help='Name of the managed instance to restore managed database to. '
+                   'This can be same managed instance, or another managed instance on same subscription. '
+                   'When not specified it defaults to source managed instance.')
+
+        c.argument('target_resource_group_name',
+                   options_list=['--dest-resource-group'],
+                   help='Name of the resource group of the managed instance to restore managed database to. '
+                   'When not specified it defaults to source resource group.')
+
+        restore_point_arg_group = 'Restore Point'
+
+        c.argument('restore_point_in_time',
+                   options_list=['--time', '-t'],
+                   arg_group=restore_point_arg_group,
+                   required=True,
+                   help='The point in time of the source database that will be restored to create the'
+                   ' new database. Must be greater than or equal to the source database\'s'
+                   ' earliestRestoreDate value. Time should be in following format: "YYYY-MM-DDTHH:MM:SS"')
+
+    with self.argument_context('sql midb list') as c:
+        c.argument('managed_instance_name', id_part=None)