+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_validators.py
@@ -123,3 +123,16 @@ def validate_subnet(cmd, namespace):
     else:
         raise CLIError('incorrect usage: [--subnet ID | --subnet NAME --vnet-name NAME]')
     delattr(namespace, 'vnet_name')
+
+
+###############################################
+#                sql managed instance         #
+###############################################
+
+
+def validate_managed_instance_storage_size(namespace):
+    # Validate if entered storage size value is an increment of 32 if provided
+    if (not namespace.storage_size_in_gb) or (namespace.storage_size_in_gb and namespace.storage_size_in_gb % 32 == 0):
+        pass
+    else:
+        raise CLIError('incorrect usage: --storage must be specified in increments of 32 GB')